Fix tests for encoding/decoding between Kotlin and Swift implementations
