+++ iosApp/iosAppTests/KMMSerializationTests.swift
@@ -36,7 +36,7 @@ import XCTest
 import JsonModel
 import KotlinModel
 
-class KMMSerializationTests: XCTestCase {
+class KMMSerializationTests: ResultSerializationTestCase {
     
     func testSurveyACodable() {
         let factory = AssessmentFactory()
@@ -147,7 +147,6 @@ class KMMSerializationTests: XCTestCase {
     func testResultCodable() {
         let factory = AssessmentFactory()
         let encoder = factory.createJSONEncoder()
-        let decoder = factory.createJSONDecoder()
 
         do {
             let json = try encoder.encode(swiftAssessmentResult)
@@ -166,47 +165,12 @@ class KMMSerializationTests: XCTestCase {
             let kmmEncoder = KotlinModel.ResultEncoder(result: kmmAssessmentResult)
             let kmmJsonString = try kmmEncoder.encodeObject()
             let kmmJson = kmmJsonString.data(using: .utf8)!
-            let decodedResult = try decoder.decode(AssessmentResultObject.self, from: kmmJson)
-            
-            checkResults(swiftResult: swiftAssessmentResult, decodedResult: decodedResult)
+            try checkAssessmentResult(from: kmmJson)
 
         } catch {
             XCTFail("Failed to encode or decode the assessment: \(error)")
         }
     }
     
-    func checkResults(swiftResult: JsonModel.ResultData, decodedResult: JsonModel.ResultData) {
-        XCTAssertEqual(swiftResult.typeName, decodedResult.typeName, "\(swiftResult.identifier)")
-        XCTAssertEqual(swiftResult.identifier, decodedResult.identifier, "\(swiftResult.identifier)")
-        XCTAssertEqual(swiftResult.startDate.timeIntervalSinceReferenceDate, decodedResult.startDate.timeIntervalSinceReferenceDate, accuracy: 0.1, "\(swiftResult.identifier)")
-        XCTAssertEqual(swiftResult.endDate.timeIntervalSinceReferenceDate, decodedResult.endDate.timeIntervalSinceReferenceDate, accuracy: 0.1, "\(swiftResult.identifier)")
-        
-        if let answerResult = swiftResult as? JsonModel.AnswerResultObject {
-            if let decoded = decodedResult as? JsonModel.AnswerResultObject {
-                XCTAssertEqual(answerResult.questionText, decoded.questionText, "\(swiftResult.identifier)")
-                XCTAssertEqual(answerResult.questionData, decoded.questionData, "\(swiftResult.identifier)")
-                XCTAssertEqual(answerResult.jsonValue, decoded.jsonValue, "\(swiftResult.identifier)")
-                XCTAssertEqual(answerResult.jsonAnswerType?.typeName, decoded.jsonAnswerType?.typeName, "\(swiftResult.identifier)")
-            }
-            else {
-                XCTFail("Failed to decode the correct type. expected=\(swiftResult), actual=\(decodedResult)")
-            }
-        }
-        else if let branchResult = swiftResult as? JsonModel.BranchNodeResult {
-            if let decoded = decodedResult as? JsonModel.BranchNodeResult {
-                branchResult.stepHistory.forEach { childResult in
-                    if let decodedChild = decoded.stepHistory.first(where: { $0.identifier == childResult.identifier }) {
-                        checkResults(swiftResult: childResult, decodedResult: decodedChild)
-                    }
-                    else {
-                        XCTFail("Failed to decode the matching child. expected=\(childResult)")
-                    }
-                }
-            }
-            else {
-                XCTFail("Failed to decode the correct type. expected=\(swiftResult), actual=\(decodedResult)")
-            }
-        }
-    }
 }
